Add more filter functionality
This commit introduced new types of concepts, views, functions, and modifiers. Those
new things were:

- Filter
- Menu
- Picker

Using the things learned above the following changes were made:

- Function called filter was made with a parameter of PredatorType in the predator class file
- PredatorType was taken out of the ApexPredators struct and place placed underneath it allowing the enum to be
used all over the project without having to type ApexPredator.PredatorType
- We used the the menu view in a new toolbar item to make a menu of the the different Dino types (land, air, sea)
- We used picker view within the menu view to make the different items in our view tappable and selectable
- We also added an ‘all’ type to our predatorType enum so when we select all in our filter menu it will show all
- In our filter function we add a conditional that says if type is .all we dont filter anything
- When something it filtered the way swift logically handles that is it creates a temporary list of items and then
waits for us to tell it where to put that temporary list. Originally we were replacing the original list with the
temporary list but when we do that the new main list becomes the filtered list meaning when we run filter for air
itll show all air but the list of all air Dinos is our new main list so if we try to filter again itll just look
through the list of air Dinos and if we chose land for example it will find no land Dinos in the list of air Dinos
so itll make the master list empty and itll just stay like that.
- To solve this we added a new collection of our apex predators in our predators class called allApexPredators that
always have the complete collections of our predators so when we filter we filter from the allApexPredators collection
and store it in our apexPredators collection and if we change it back to ‘all’ in our filter menu we just set
apexPredators = allApexPredators